refactor(gastown): extract scheduling module, parallelize alarm loop, fix refinery recovery
Extract dispatchAgent, dispatchUnblockedBeads, schedulePendingWork, and
hasActiveWork into town/scheduling.ts (~360 lines out of Town.do.ts).

Improve refinery recovery after container restart:
- Reorder alarm: patrols run BEFORE scheduling so zombie agents
  detected by witnessPatrol are dispatched in the same tick instead
  of waiting for the next one.
- Exclude refineries from schedulePendingWork so they always go
  through processReviewQueue with the full system prompt (branch,
  strategy, gates). recoverStuckReviews resets the MR bead to open
  after the timeout, and processReviewQueue re-pops it correctly.
- Reduce REVIEW_RUNNING_TIMEOUT_MS from 5 min to 2.5 min so the
  review queue recovers faster after a container restart.

Parallelize the alarm loop for lower latency:
- Phase 1: witnessPatrol + deaconPatrol run in parallel (patrols)
- Phase 2: review pipeline (sequential) runs in parallel with
  schedulePendingWork (disjoint agent types)
- Phase 3: mail delivery, nudge expiry, re-escalation, and triage
  dispatch all run in parallel (fully independent)
- Zombie detection container status checks run in parallel via
  Promise.allSettled instead of sequential for-loop.
